Traces extraction API

Extract activity channel traces

ExtractRegisteredData.extract_activity_am_reg — Function

Extracts activity marker activity from camera-alignment registration. Returns any errors encountered.

Arguments

param_path::Dict: Dictionary of paths to relevant files.

param::Dict: Directory of imaging parameters.

shear_params_dict::Dict: Dictionary of shear-correction parameters used in the marker channel.

crop_params_dict::Dict: Dictionary of cropping parameters used in the marker channel.

mhd_path_key::String (optional): Key in param_path to path to unprocessed activity-channel MHD files. Default path_dir_mhd

roi_dir_key::String (optional): Key in param_path to directory of neuron ROI files. Default path_dir_roi_watershed.

transform_key::String (optional): Key in param_path to file name of transform parameter files. Default path_dir_transformed_activity_marker_avg.

ExtractRegisteredData.extract_traces — Function

Extracts traces from neuron labels.

Arguments

inverted_map: Dictionary of dictionaries mapping time points to original ROIs, for each neuron label

gcamp_data_dir: Directory containing GCaMP activity data

ExtractRegisteredData.make_traces_array — Function

Turns a traces dictionary into a traces array. Also outputs a heatmap of the array, and the labels of which neurons correspond to which rows.

Arguments

traces::Dict: Dictionary of traces.

threshold::Real (optional): Minimum number of time points necessary to include a neuron in the traces array. Default 1 (all ROIs displayed). It is recommended to set either this or the valid_rois parameter.

valid_rois (optional): If set, use this set of neurons (in order) in the array.

contrast::Real (optional): If set, increases the contrast of the heatmap. Does not change the output array.

replace_blank::Bool (optional): If set, replaces all blank entries in the traces (where the neuron was not found in that time point) with the median activity for that neuron.

cluster::Bool (optional): If set to true, cluster the neurons by activity similarity.

SWF360 error rate computation

ExtractRegisteredData.error_rate — Function

Computes the error rate of the traces, assuming they are all supposed to have constant activity.

Arguments

traces: the traces

Optional keyword arguments

factor::Real: multiplicative amount by which activity must differ from median to count as an error. Default 1.5.

gfp_thresh::Real: threshold amount for a neuron to be counted as having GFP (error rate is adjusted to account for mismatched neurons that have similar activity)

num_thresh::Real: minimum amount of neurons needed to consider a trace

Visualization tools

ExtractRegisteredData.output_roi_candidates — Function

Outputs neuron ROI candidates and a plot of their activity.

Arguments

traces::Dict: Dictionary of traces for all ROIs

inv_map::Dict: Dictionary that maps ROI identity to time points and UNet ROI labels

param_path::Dict: Dictionary of paths to relevant files

param::Dict: Dictionary of parameter values

get_basename::Function: Function that gets the basename of an MHD file

channel::Integer: Channel of the image to be displayed

t_range: All time points