Registration-based neuron matching API

Commonly used functions

ExtractRegisteredData.extract_roi_overlap — Function

extract_roi_overlap(
    best_reg::Dict, param_path::Dict, param::Dict; reg_dir_key::String="path_dir_reg",
    transformed_dir_key::String="path_dir_transformed", reg_problems_key::String="path_reg_prob",
    param_path_moving::Union{Dict,Nothing}=nothing
)

Extracts ROI overlaps and activity differences.

Arguments

best_reg::Dict: Dictionary of best registration values.

param_path::Dict: Dictionary of paths to relevant files.

param::Dict: Dictionary of parameter values.

reg_dir_key::String (optional): Key in param_path corresponding to the registration directory. Default path_dir_reg.

transformed_dir_key::String (optional): Key in param_path corresponding to the transformed ROI save directory. Default path_dir_transformed.

reg_problems_key::String (optional): Key in param_path corresponding to the registration problems file. Default path_reg_prob.

param_path_moving::Union{Dict, Nothing}: If set, the param_path dictionary corresponding to the moving dataset. Otherwise, the method will assume

the moving and fixed datasets have the same path dictionary.

ExtractRegisteredData.make_regmap_matrix — Function

make_regmap_matrix(
    roi_overlaps::Dict, roi_activity_diff::Dict, q_dict::Dict, best_reg::Dict, param::Dict;
    watershed_errors::Union{Nothing,Dict}=nothing, max_fixed_t::Int=0
)

Assimilates ROIs from all time points and generates a quality-of-match matrix.

Arguments

roi_overlaps::Dict: a dictionary of dictionaries of ROI matches for each pair of moving and fixed time points

roi_activity_diff::Dict: a dictionaon of dictionaries of the difference between normalized red ROI acivity, for each pair of moving and fixed time points

q_dict::Dict: a dictionary of dictionaries of dictionaries of the registration quality for each metric for each resolution for each pair of moving and fixed time points

best_reg::Dict: a dictionary of the registration that's being used, expressed as a tuple of resolutions, for each pair of moving and fixed time points

param::Dict: a dictionary of parameter settings to use, including:

activity_diff_threshold::Real: parameter that controls how much red-channel activity differences are penalized in the matrix.

Smaller values = greater penalty. Default 0.3. * watershed_error_penalty::Real: parameter that control how much watershed/UNet errors are penalized in the matrix. Smaller values = greater penalty. Default 0.5. * quality_metric: metric to use in the q_dict parameter. Default NCC. * matrix_self_weight::Real: amount of weight in the matrix to assign each ROI to itself. Default 0.5. * size_mismatch_penalty::Real: penalty to apply to overlapping ROIs that don't fully overlap. Larger values = greater penalty. Default 2.

watershed_errors::Union{Nothing, Dict}: a dictionary of ROIs that might have had watershed or UNet errors for every time point.

This dictionary must be pre-shifted if moving and fixed datasets are not the same. Ignored if nothing.

max_fixed_t::Int: If the moving and fixed datasets are not the same, this number is added to each moving dataset time point to distinguish the datasets in the matrix and label map.

Returns

regmap_matrix, a matrix whose (i,j)th entry encodes the quality of the match between ROIs i and j.

label_map: a dictionary of dictionaries mapping original ROIs to new ROI labels, for each time point.

ExtractRegisteredData.find_neurons — Function

find_neurons(regmap_matrix, label_map::Dict, param::Dict)

Groups ROIs into neurons based on a matrix of overlaps.

Arguments:

regmap_matrix: Matrix of distances between the ROIs

label_map::Dict: Dictionary of dictionaries mapping original ROIs to new ROI labels, for each time point.

param::Dict: Dictionary containing cluster_overlap_thresh and cluster_height_thresh parameter settings to use for clustering.

Returns

new_label_map: Dictionary of dictionaries mapping original ROIs to neuron labels, for each time point.

inv_map: Dictionary of dictionaries mapping time points to original ROIs, for each neuron label

hmer: Raw clusters of the dataset.

ExtractRegisteredData.register_immobilized_rois — Function

register_immobilized_rois(regmap_matrix, label_map_regmap, inv_map_regmap, label_map_freelymoving, valid_rois_freelymoving, param, reg_timept)

Registers immobilized ROIs to the freely moving dataset.

Arguments:

regmap_matrix: Matrix of distances between the ROIs

label_map_regmap::Dict: Dictionary of dictionaries mapping original ROIs to new ROI labels for the freely-moving-to-immobilized registration.

inv_map_regmap::Dict: Dictionary of dictionaries mapping new ROI labels to original ROIs, for the freely-moving-to-immobilized registration.

label_map_freelymoving::Dict: Dictionary of dictionaries mapping original ROIs to new ROI labels in the freely-moving dataset (one of the outputs of find_neurons).

valid_rois_freelymoving: List of which ROIs correspond to neurons in the freely-moving dataset.

param::Dict: Dictionary containing max_t parameter setting to use for registration.

reg_timept: Time point of the immobilized dataset to register.

Returns:

roi_matches: Dictionary of dictionaries mapping immobilized ROIs to matched freely moving ROIs.

inv_matches: Dictionary of dictionaries mapping matched freely moving ROIs to immobilized ROIs.

roi_matches_best: Dictionary of dictionaries mapping immobilized ROIs to matched freely moving ROIs, with only the best match for each immobilized ROI.

roi_match_confidence: Dictionary of dictionaries mapping immobilized ROIs to the confidence of the match.

ExtractRegisteredData.invert_label_map — Function

invert_label_map(label_map)

Inverts ROI label map label_map, so that new ROIs map back to dictionaries mapping original time points to original ROIs.

transformix helper functions

ExtractRegisteredData.run_transformix_centroids — Function

run_transformix_centroids(path, output, centroids, parameters, transformix_dir)

Runs transformix to map centroids from the fixed volume to the moving volume. Note that this is the opposite of run_transformix_roi, because of the way transformix is implemented.

Arguments

path::String: Path to directory containing elastix registration

output::String: Output file directory

centroids: File containing centroids to transform

parameters::String: Transform parameter file

transformix_dir::String: Path to transformix executable

ExtractRegisteredData.run_transformix_roi — Function

run_transformix_roi(path::String, input::String, output::String, parameters::String, parameters_roi::String, transformix_dir::String)

Runs transformix to map an ROI image from the moving volume to the fixed volume. Modifies various transform parameters to force nearest-neighbors interpolation. Returns the resulting transformed image. Note that this is the opposite of run_transformix_centroids, because of the way transformix is implemented.

Arguments

path::String: Path to directory containing elastix registration. Other paths should be absolute; they are NOT relative to this.

input::String: Input file path

output::String: Output file directory path

parameters::String: Path to transform parameter file

parameters_roi::String: Filename to save modified parameters.

transformix_dir::String: Path to transformix executable

ExtractRegisteredData.run_transformix_img — Function

run_transformix_img(path::String, output::String, input::String, parameters::String, parameters_roi::String, transformix_dir::String; interpolation_degree::Int=1)

Runs transformix to map an image from the moving volume to the fixed volume. Returns the resulting transformed image. Note that this is the opposite of run_transformix_centroids, because of the way transformix is implemented.

Arguments

path::String: Path to directory containing elastix registration

output::String: Output file directory

input::String: Input file path

parameters::String: Transform parameter file

transformix_dir::String: Path to transformix executable

interpolation_degree::Int (optional): Interpolation degree. Default 1 (linear interpolation).

Other helper functions

ExtractRegisteredData.pairwise_dist — Function

pairwise_dist(regmap_matrix; threshold::Real=1e-16, dtype::Type=Float64)

Finds the distances between each pair of ROIs based on similarity between rows of the regmap_matrix.

Arguments

regmap_matrix: matrix of registration matches

Optional keyword arguments

threshold::Real: a small number added to the denominator to ensure there are no divide-by-zero errors. Default 1e-16.

dtype::Type: Data type of matrix. Default Float64.

ExtractRegisteredData.register_neurons_overlap — Function

register_neurons_overlap(roi_inferred, roi, activity_inferred, activity)

Matches ROIs from registered time points together based on the overlap heuristic.

Arguments

inferred_roi: ROI image array from the moving volume that has had the registration applied to it (eg via transformix)

roi: ROI image array from the fixed time point

inferred_activity: array of activities of each ROI in the marker channel from the moving time point

activity: array of activities of each ROI in the marker channel from the fixed time point

Returns

overlapping_rois: a dictionary of ROI matches between the moving and fixed volumes. The values of the dictionary are the amount of overlap as a fraction of total ROI size. Eg: if ROI 15 in the moving volume has size 100, ROI 10 in the fixed volume has size 80, and the overlap has size 50, (15, 10) => (0.5, 0.625) would be an entry in the dictionary.

activity_mismatch: a dictionary of the differences between the normalized activity of the ROIs between the moving and fixed time points.

ExtractRegisteredData.update_label_map — Function

update_label_map(label_map, matches)

Updates ROI label map label_map to include ROI matches matches, and returns updated version.

ExtractRegisteredData.delete_smeared_neurons — Function

delete_smeared_neurons(roi, threshold)

Deletes neurons that might come from huge deformations in registration from an roi (neurons that have more pixels than threshold).

ExtractRegisteredData.merge_confocal_data! — Function

merge_confocal_data!(combined_data_dict::Dict, data_dict::Dict, data_dict_2::Dict, dataset::String; traces_key="traces_array", zscored_traces_key="raw_zscored_traces_array", F_F20_key="traces_array_F_F20")

This function merges neuron identities from two datasets from the same animal, data_dict and data_dict_2, into a new dictionary combined_data_dict. The function is used to combine data from two different datasets.

Arguments

combined_data_dict::Dict: A dictionary that will contain the merged data.

data_dict::Dict: A dictionary containing the first dataset to be merged.

data_dict_2::Dict: A dictionary containing the second dataset to be merged.

dataset::String: A string representing the name of the second dataset in the first dataset's dictionray.

traces_key::String: A string representing the key for the traces array in the dictionaries. Default is "traces_array".

zscored_traces_key::String: A string representing the key for the zscored traces array in the dictionaries. Default is "rawzscoredtraces_array".

F_F20_key::String: A string representing the key for the F/F0 traces array in the dictionaries. Default is "tracesarrayF_F20".

The function returns nothing, but modifies the combined_data_dict dictionary in place.

ExtractRegisteredData.match_neurons_across_datasets — Function

match_neurons_across_datasets(label_map_1::Dict, label_map_2::Dict, inv_map_reg::Dict, max_fixed_t::Int)

Matches neurons across multiple datasets.

Arguments

label_map_1::Dict: Label map of ROI/time points to neuron ID for dataset 1.

label_map_2::Dict: Label map of ROI/time points to neuron ID for dataset 2.

inv_map-reg::Dict: Map of neuron ID to ROI/time points for the registration between datasets. Time points in dataset 2 are shifted by max_fixed_t.

max_fixed_t::Int: Maximum time point in the first dataset.